Refit an ARIMA Model on a Subset of the Time Series

There seems to be no function in package forecast (as of version 8.2) to re-estimate an ARIMA model on a subset of the original time series. This update method does exactly that.

# S3 method for Arima
update(object, subset, ...)

Arguments

object
an object of class "Arima", e.g., from forecast::auto.arima().

subset
an integer vector selecting part of the original time series (and external regressors).

...
further arguments to be passed to arima().

Value

the updated model.